+++ Precog/src/precog/Pt_Avg_Perc_Calc.java
@@ -40,7 +40,7 @@ public class Pt_Avg_Perc_Calc implements Runnable
 		for (int i = start_idx; i <= end_idx; i++)
 		{
 			long posBoard = board | pos_river_set[i];
-			double myRating = Precog.rate(Precog.getHighestHand(hand, posBoard));
+			int myRating = Precog.rate(Precog.getHighestHand(hand, posBoard));
 			
 			int totalNum = 0;
 		    double notBigger = 0.d;
